requirements/production.txt
django==1.6
djrill==0.7.0
django-extensions==1.2.5
django-braces==1.2.2
django-floppyforms>=1.1
django-crispy-forms>=1.4.0